refactor: split message schema modules
